+++ lib/game/chunk/utils.h
@@ -4,7 +4,7 @@
 
 namespace Game {
 
-int calculate_ref_count(const std::vector<long long> &data,
+int calculate_ref_count(const std::vector<int64_t> &data,
                         const std::vector<int> &palette);
 
 }
